modification

+++ routes/web.php
@@ -18,42 +18,41 @@ use Illuminate\Support\Facades\Route;
 |
 */
 
-$router->get('/fd', function () use ($router) {
+$router->get('/test', function () use ($router) {
     return $router->app->version();
 });
-
-$router->get('fd', ['middleware' => '', function () use ($router) {
-    return $router->app->version();
-}]);
 
 $router->group([], function () use ($router) {
     $router->post('/login', ['uses' => 'AuthController@login']);
     $router->post('/register', ['uses' => 'AuthController@register']);
+    $router->post('/logout', ['uses' => 'AuthController@logout']);
+    Route::get('/me', ['uses' => 'AuthController@show']);
 });
 
 Route::group(['prefix' => 'users'], function () {
-    Route::get('/', ['uses' => 'UserController@index']);
-    Route::post('/', ['uses' => 'UserController@store']);
+    Route::get('/', ['uses' => 'UserController@index', 'middleware' => ['auth:admin']]);
+    Route::get('/{user}', ['uses' => 'UserController@show', 'middleware' => ['auth:admin']]);
+    Route::post('/', ['uses' => 'UserController@store', 'middleware' => ['auth:admin']]);
     Route::put('/{user}', ['uses' => 'UserController@update']);
-    Route::delete('/{user}', ['uses' => 'UserController@destroy']);
+    Route::delete('/{user}', ['uses' => 'UserController@destroy', 'middleware' => ['auth:admin']]);
 });
 
 
 Route::group(['prefix' => 'servers'], function () {
     Route::get('/', ['uses' => 'ServerController@index']);
-    Route::post('/', ['uses' => 'ServerController@store']);
-    Route::put('/{server}', ['uses' => 'ServerController@update']);
-    Route::delete('/{server}', ['uses' => 'ServerController@destroy']);
+    Route::post('/', ['uses' => 'ServerController@store', 'middleware' => ['auth:admin']]);
+    Route::put('/{server}', ['uses' => 'ServerController@update', 'middleware' => ['auth:admin']]);
+    Route::delete('/{server}', ['uses' => 'ServerController@destroy', 'middleware' => ['auth:admin']]);
 });
 
 Route::group(['prefix' => 'services'], function () {
     Route::get('/', ['uses' => 'ServiceController@index']);
-    Route::post('/', ['uses' => 'ServiceController@store']);
-    Route::put('/{service}', ['uses' => 'ServiceController@update']);
-    Route::delete('/{service}', ['uses' => 'ServiceController@destroy']);
+    Route::post('/', ['uses' => 'ServiceController@store', 'middleware' => ['auth:admin']]);
+    Route::put('/{service}', ['uses' => 'ServiceController@update', 'middleware' => ['auth:admin']]);
+    Route::delete('/{service}', ['uses' => 'ServiceController@destroy', 'middleware' => ['auth:admin']]);
 });
 
-Route::group(['prefix' => 'service_user'], function () {
+Route::group(['prefix' => 'service_user', 'middleware' => ['auth:admin']], function () {
     Route::get('/', ['uses' => 'UserServiceController@index']);
     Route::post('/', ['uses' => 'UserServiceController@store']);
     Route::put('/{userService}', ['uses' => 'UserServiceController@update']);
@@ -63,8 +62,8 @@ Route::group(['prefix' => 'service_user'], function () {
 
 
 Route::group(['prefix' => 'connections'], function () {
-    Route::get('/', ['uses' => 'ConnectionController@index']);
+    Route::get('/', ['uses' => 'ConnectionController@index', 'middleware' => ['auth:admin']]);
     Route::post('/', ['uses' => 'ConnectionController@store']);
-    Route::put('/{connection}', ['uses' => 'ConnectionController@update']);
-    Route::delete('/{connection}', ['uses' => 'ConnectionController@destroy']);
+    Route::put('/{connection}', ['uses' => 'ConnectionController@update', 'middleware' => ['auth:admin']]);
+    Route::delete('/{connection}', ['uses' => 'ConnectionController@destroy', 'middleware' => ['auth:admin']]);
 });
